# spec/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Api Documentation",
    "version" : "1.0",
    "title" : "Api Documentation",
    "termsOfService" : "urn:tos",
    "contact" : { },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host" : "localhost",
  "basePath" : "/",
...
  "paths" : {
...
    "/api/customers/list" : {
      "get" : {
        "tags" : [ "customer-controller" ],
        "summary" : "fetchAllCustomers",
        "operationId" : "fetchAllCustomersUsingGET",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "customerType",
          "in" : "query",
          "description" : "customerType",
          "required" : false,
          "type" : "string",
          "default" : "Betaler"
        }, {
          "name" : "expand",
          "in" : "query",
          "description" : "expand",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "parentCompany",
          "in" : "query",
          "description" : "parentCompany",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "skipToken",
          "in" : "query",
          "description" : "skipToken",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/CustomerDTO"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
    "/api/v1/terms/customer/create" : {
      "post" : {
        "tags" : [ "customer-terms-controller" ],
        "summary" : "create",
        "operationId" : "createUsingPOST",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "customerTermsDTO",
          "description" : "customerTermsDTO",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/CustomerTermsDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/CustomerTermsDTO"
            }
          },
          "201" : {
            "description" : "Created"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/v1/terms/customer/list" : {
      "get" : {
        "tags" : [ "customer-terms-controller" ],
        "summary" : "list",
        "operationId" : "listUsingGET_1",
        "produces" : [ "*/*" ],
        "parameters" : [ {
          "name" : "customerNumber",
          "in" : "query",
          "description" : "customerNumber",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "salesOffice",
          "in" : "query",
          "description" : "salesOffice",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/CustomerTermsDTO"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/v1/terms/customer/list/active" : {
      "get" : {
        "tags" : [ "customer-terms-controller" ],
        "summary" : "listAllActive",
        "operationId" : "listAllActiveUsingGET",
        "produces" : [ "*/*" ],
        "parameters" : [ {
          "name" : "customerNumber",
          "in" : "query",
          "description" : "customerNumber",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "salesOffice",
          "in" : "query",
          "description" : "salesOffice",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/CustomerTermsDTO"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/v1/terms/customer/save/{id}" : {
      "put" : {
        "tags" : [ "customer-terms-controller" ],
        "summary" : "save",
        "operationId" : "saveUsingPUT",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "customerTermsDTO",
          "description" : "customerTermsDTO",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/CustomerTermsDTO"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/CustomerTermsDTO"
            }
          },
          "201" : {
            "description" : "Created"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "ContactPersonDTO" : {
      "type" : "object",
      "properties" : {
        "emailAddress" : {
          "type" : "string"
        },
        "firstName" : {
          "type" : "string"
        },
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "lastName" : {
          "type" : "string"
        },
        "mobileNumber" : {
          "type" : "string"
        }
      },
      "title" : "ContactPersonDTO"
    },
    "CustomerBranchDTO" : {
      "type" : "object",
      "properties" : {
        "branch" : {
          "type" : "string"
        },
        "branchSystem" : {
          "type" : "string"
        },
        "branchSystemText" : {
          "type" : "string"
        },
        "branchText" : {
          "type" : "string"
        }
      },
      "title" : "CustomerBranchDTO"
    },
    "CustomerDTO" : {
      "type" : "object",
      "properties" : {
        "abcClassification" : {
          "type" : "string"
        },
        "businessPartnerCategory" : {
          "type" : "string"
        },
        "changedDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "changedTime" : {
          "type" : "string"
        },
        "city" : {
          "type" : "string"
        },
        "company" : {
          "type" : "string"
        },
        "contactPersons" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ContactPersonDTO"
          }
        },
        "country" : {
          "type" : "string"
        },
        "creditScore" : {
          "type" : "string"
        },
        "customerBranchList" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/CustomerBranchDTO"
          }
        },
        "customerGroup" : {
          "type" : "string"
        },
        "customerName" : {
          "type" : "string"
        },
        "customerNumber" : {
          "type" : "string"
        },
        "customerType" : {
          "type" : "string"
        },
        "distributionChannel" : {
          "type" : "string"
        },
        "division" : {
          "type" : "string"
        },
        "firmPostalNumber" : {
          "type" : "string"
        },
        "hasCentralAgreement" : {
          "type" : "string"
        },
        "houseNumber" : {
          "type" : "string"
        },
        "isConstructionCustomer" : {
          "type" : "string"
        },
        "motherCompany" : {
          "type" : "string"
        },
        "name1" : {
          "type" : "string"
        },
        "name2" : {
          "type" : "string"
        },
        "nodeCustomerList" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/NodeCustomerDTO"
          }
        },
        "organizationNumber" : {
          "type" : "string"
        },
        "payer" : {
          "type" : "string"
        },
        "permissionHazardousWaste" : {
          "type" : "string"
        },
        "postBox" : {
          "type" : "string"
        },
        "postBoxNumber" : {
          "type" : "string"
        },
        "postalNumber" : {
          "type" : "string"
        },
        "regStrGrp" : {
          "type" : "string"
        },
        "region" : {
          "type" : "string"
        },
        "streetAddress" : {
          "type" : "string"
        },
        "timeZone" : {
          "type" : "string"
        }
      },
      "title" : "CustomerDTO"
    },
    "CustomerTermsDTO" : {
      "type" : "object",
      "properties" : {
        "additionForAdminFee" : {
          "type" : "boolean"
        },
        "additionForRoadTax" : {
          "type" : "boolean"
        },
        "agreementEndDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "agreementStartDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "comment" : {
          "type" : "string"
        },
        "contractTerm" : {
          "type" : "string"
        },
        "createdBy" : {
          "type" : "string"
        },
        "createdDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "customerName" : {
          "type" : "string"
        },
        "customerNumber" : {
          "type" : "string"
        },
        "customerRequireNotification" : {
          "type" : "boolean"
        },
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "indexRent" : {
          "type" : "string"
        },
        "indexTransport" : {
          "type" : "string"
        },
        "indexWaste" : {
          "type" : "string"
        },
        "invoiceInterval" : {
          "type" : "string"
        },
        "lastModifiedBy" : {
          "type" : "string"
        },
        "lastModifiedDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "level" : {
          "type" : "string"
        },
        "metalPricing" : {
          "type" : "string"
        },
        "metalSetDateForOffer" : {
          "type" : "string",
          "format" : "date-time"
        },
        "monthsToFreezeCustomerFromAdjustments" : {
          "type" : "integer",
          "format" : "int32"
        },
        "newOrEstablishedCustomer" : {
          "type" : "string"
        },
        "nodeNr" : {
          "type" : "string"
        },
        "notificationMailAddress" : {
          "type" : "string"
        },
        "number" : {
          "type" : "integer",
          "format" : "int32"
        },
        "paymentCondition" : {
          "type" : "string"
        },
        "priceAdjustmentDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "quarterlyAdjustment" : {
          "type" : "string"
        },
        "region" : {
          "type" : "string"
        },
        "requiredNumberOfDaysNotice" : {
          "type" : "integer",
          "format" : "int32"
        },
        "salesOffice" : {
          "type" : "string"
        },
        "source" : {
          "type" : "string"
        },
        "specialConditionAction" : {
          "type" : "string"
        },
        "year" : {
          "type" : "integer",
          "format" : "int32"
        }
      },
      "title" : "CustomerTermsDTO"
    },
...
    "NodeCustomerDTO" : {
      "type" : "object",
      "properties" : {
        "hkunnr" : {
          "type" : "string"
        },
        "kunnr" : {
          "type" : "string"
        }
      },
      "title" : "NodeCustomerDTO"
    },
...
  }
}